Add files via upload
Simulates battles for dice in the game Risk Global Domination. The rules are that in order for an attacker to win, their die has to be larger than the defender die. In addition, an attacker can attack with up to 3 dice while a defender can only defend with up to 2 dice at a time.